Refactor: using Separation of Concept principle to separate DOM manipulation out of the index.js
