+++ index.html
@@ -1,26 +1,44 @@
 <!-- index.html file -->
 <!DOCTYPE html>
 <html lang="pt-br">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="icon" type="image/png" href="img/favicon.ico">
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;700&display=swap">
-    <link rel="stylesheet" href="css/styles.css">
-    <!-- Game precisa ser referenciado primeiro, pq ele é usado pelo index. A ORDEM IMPORTA AQUI -->
-    <script src="https://unpkg.com/dexie/dist/dexie.js"></script>
-    <script type="module" src="js/enums/RefreshTypeEnum.js"></script> 
-    <script type="module" src="js/utils/dateUtils.js"></script> 
-    <script type="module" src="js/utils/colorUtils.js"></script> 
-    <script type="module" src="js/utils/validationUtils.js"></script> 
-    <script type="module" src="js/database/dbInit.js"></script> 
-    <script type="module" src="js/database/gameDB.js"></script> 
-    <script type="module" src="js/database/taskDB.js"></script> 
-    <script type="module" src="js/data/Game.js"></script> 
-    <script type="module" src="js/data/Task.js"></script> 
-    <script type="module" src="js/index.js"></script> 
-    <title>Gacha Management</title>
-</head>
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <link rel="icon" type="image/png" href="img/favicon.ico">
+        <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;700&display=swap">
+        <link rel="stylesheet" href="css/styles.css">
+        
+        <!-- External Library -->
+        <script src="https://unpkg.com/dexie/dist/dexie.js"></script>
+    
+        <!-- Utility Scripts -->
+        <script type="module" src="js/enums/RefreshTypeEnum.js"></script> 
+        <script type="module" src="js/utils/dateUtils.js"></script> 
+        <script type="module" src="js/utils/colorUtils.js"></script> 
+        <script type="module" src="js/utils/validationUtils.js"></script> 
+    
+        <!-- Database Initialization -->
+        <script type="module" src="js/database/dbInit.js"></script> 
+        <script type="module" src="js/database/gameDB.js"></script> 
+        <script type="module" src="js/database/taskDB.js"></script> 
+    
+        <!-- Data Models -->
+        <script type="module" src="js/data/Game.js"></script> 
+        <script type="module" src="js/data/Task.js"></script> 
+    
+        <!-- UI Scripts (Newly Added) -->
+        <script type="module" src="js/ui/gameUI.js"></script>
+        <script type="module" src="js/ui/taskUI.js"></script>
+        <script type="module" src="js/ui/formHandler.js"></script>
+        <script type="module" src="js/ui/dropdownHandler.js"></script>
+        <script type="module" src="js/ui/inputValidation.js"></script>
+    
+        <!-- Main Script -->
+        <script type="module" src="js/index.js"></script> 
+    
+        <title>Gacha Management</title>
+    </head>
+    
 <body>
     <section>
         <h1 class="main-title">
