+++ js/utils/colorUtils.js
@@ -1,5 +1,5 @@
 // colorUtils.js
-export default function getRandomColor() {
+export function getRandomColor() {
     const colors = [
         "#d9d9d9", // Cinza Claro
         "#f2e6d9", // Bege Claro
